metrics: Reduce size of JSON output
Change the representation of the histogram in the JSON output metrics
to only contain the index and value of non-zero buckets. Since the
bucket boundaries are determinstic, based on histogram parameters,
these can be reconstructed rather than stored. Further, the histogram
is represented in column-order rather than row-order, i.e., each field
is represented in its own vector, rather than having a vector of
buckets. This removes the need to repeat the field names for every
bucket in the JSON output.

The combination of these changes leads to a size reduction of 4-5x in
total uncompressed JSON output.