feat: add keyspace TTL support
Adds support for specifying a keyspace-level TTL. This applies to
storage commands and hash and sorted set increment commands.

For Momento, this is implemented by providing the TTL and choosing
no refresh for existing collections.

For Redis, this is implemented by fusing an EXPIRE/PEXPIRE with NX
option onto the equivalent commands.

For Memcache, we provide the TTL for storage commands.