fix: use specific seeds to initialize PRNGs
Previously, we were using `seed_from_u64(0)` to initialize PRNGs for
key and topic generation. This yields a predictable set of keys for
a fixed config, however, the ordering of requests within the keyspace
was random per-run as the PRNG for each workload generating thread was
seeded from the OS entropy pool.

This change moves to using fixed random seeds to generate the keys and
topics. We use a separate fixed random seed to initialize a PRNG which
is then used to bootstrap each workload generating thread's own PRNG.

The result is that configurations that use a single workload generator
thread should produce more deterministic workloads.