Add debian packaging setup (#25)
This commit introduces a (mostly) idiomatic debian packaging script for
rpc-perf. Of course, it would not be enough to get into any of the
official debian repos but as a setup for creating APT packages for
custom repos it should work just fine.

In order to build a package you can do the following

    dpkg-buildpackage -b

This will build a rpc-perf deb as well as a rpc-perf-dbgsym deb (or
equivalent) and place them in the parent directory of the project.
Unfortunately, it is quite difficult to change the directory the debs
are placed in in a way that is compatible with all the versions of dpkg
and debhelper that it is necessary to use if you want to support all
versions of both debian and ubuntu.

Note that if you don't have signing keys you will likely need to do

    dpkg-buildpackage -b --no-sign

or else you will get an error.

Making a New Release
====================
In order to make a new release you will need to update the
debian/changelog file with a new entry for the new release number. I am
not sure how well the -alpha.0 suffix works with the changelog format
but it seems to work as if it is the release number when I do it.

Note that the changelog format is rather finicky. Messing up the number
of spaces of indentation before each of the items will cause
dpkg-buildpackage to error. To see the docs for how to write the
changelog entry see
https://www.debian.org/doc/manuals/maint-guide/dreq.en.html#changelog

Detailed Notes
==============
- I have set things up to use debhelper compat version 10. Compat
  version 10 should work across all currently supported versions of
  debian and ubuntu.
- This will use the version of rust installed on your system and so is
  not hermetic. Not a problem for most use cases but would make
  upstreaming this into debian hard.
- By default debhelper splits the package into two packages: rezolus and
  rezolus-dbgsym (or -dbg or .ddeb). The rezolus package contains no
  debug symbols while the dbgsym package contains them separately.