use config interval for histogram integration period
Previously, we were updating the histogram snapshots on a secondly
basis. This means that the percentiles were across the "previous"
second. This assumes that secondly metrics would be used and
provides the highest resolution view.

However, when metrics are not collected every second, data is lost
as some periods would not be reflected in the collected metrics at
all.

To address this, we refresh the histogram snapshots on an interval
that matches the CLI output interval. This allows the user to set
an interval that makes sense for their environment.