feat: add internal ratelimit controller (#144)
The configuration currently allows rpc-perf to be setup with a single static
ratelimit for the entire run and exposes an endpoint through which this can
be controlled externally. This adds a basic workload definition for dynamic
ratelimiting in the configuration, as well as a background thread that modifies
the ratelimit using the endpoint so that rpc-perf can run with a dynamic
ratelimit without relying on an external controller.