config: move metrics configuration to separate section
Move configuration of metrics output to a separate section
in the configuration file. At the moment, allow metrics
configuration to be specified in the old, i.e, via the
general section, or the new, i.e., via the metrics section.
If configuration parameters in both sections are in conflict,
use the ones from the metrics section.