deps: update histogram, metriken, and ringlog (#229)
Update dependencies to move towards a single version
of histogram (0.11.0) and metriken (0.7.0). This is
a breaking change in the interface of calls to percentile
functions, which can now return null when the histogram
is empty.

The stats functions call the percentiles and publish
them using an iterator over the returned vector. Rather
than changing the interface of the wrapper function to
returning null when the histogram is empty, as the histogram
does, it returns an empty vector to metrics callers to
preserve the existing code. Since the empty vector is
iterated over, none of the percentile metrics are published.