momento http performance improvement (#320)
Improves the performance of the Momento HTTP client by increasing
the http2 initial window size for stream and connection as well as
increasing the max frame size. This reduces the number of frames
required for large payloads and results in a large performance
improvement that reduces asymmetry between read and write
performance.

Additionally, we add a time-to-first-byte metric for cache client,
as this was useful in understanding this issue.