fix: momento cache h2 client high write latency (#344)
We saw poor latencies when testing Momento Cache API over HTTP2. By using a minimal client outside of this repo, we were able to identify some core changes to how we handle both GETs and PUTs by moving the chunking logic into our code and relying less on the h2 crate.

This PR is believed to address poor PUT performance and increases the reliability of the TTFB metrics.