Add MomentoProtosocket cache protocol
MomentoProtosocket is an incubating connection-oriented protocol
using [protosocket-rpc](https://github.com/kvc0/protosocket/). It
further integrates with a new `/endpoints` http+json api on Momento
servers, which return a list of socket addresses that speak this
new protocol directly.

As with all Momento protocols, the protosocket flavor uses TLS.
It is expected that momento protosocket will show higher throughput,
lower latency, and lower network traffic than grpc or bare http
calls. That theory can be tested with rpc-perf.

I used a FuturesUnordered for concurrency, and a task per client
rather than a task per concurrency. I think that should play a little
more nicely with the runtime when using few connections and high
concurrency.

Note that it is expected that users will actually only use 1 client.
The new protosocket client includes a connection pool internally.
But for rpc perf metrics, I'm making N clients, each with 1 connection.
It shouldn't make much of a difference in practice, but it's a small
delta from reality.